+++ codemp/rd-gwz/tr_ghoul2.cpp
@@ -741,7 +741,7 @@ public:
 	int				surfaceNum;
 	surfaceInfo_v	&rootSList;
 	shader_t		*cust_shader;
-	int64_t			fogNum;
+	int				fogNum;
 	qboolean		personalModel;
 	CBoneCache		*boneCache;
 	int				renderfx;
